refactor: Header переименован в Navbar
