+++ src/components/common/navbar/index.ts
@@ -1,0 +1,1 @@
+export * from './navbar';
