+++ src/components/common/navbar/navbar.module.css
@@ -8,7 +8,7 @@
   content: '';
 }
 
-.header {
+.navbar {
   display: flex;
   justify-content: space-between;
   border-bottom: 1px solid var(--black);
